+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Era.hs
@@ -11,28 +11,25 @@ module Cardano.Ledger.Api.Era (
   ByronEra,
 
   -- ** Shelley
-  Shelley,
   ShelleyEra,
 
   -- ** Allegra
-  Allegra,
   AllegraEra,
 
   -- ** Mary
-  Mary,
   MaryEra,
 
   -- ** Alonzo
-  Alonzo,
   AlonzoEra,
 
   -- ** Babbage
-  Babbage,
   BabbageEra,
 
   -- ** Conway
-  Conway,
   ConwayEra,
+
+  -- ** Dijkstra
+  DijkstraEra,
 
   -- ** Latest Known
   LatestKnownEra,
@@ -54,10 +51,10 @@ module Cardano.Ledger.Api.Era (
   atMostEra,
 ) where
 
-import Cardano.Ledger.Allegra (Allegra, AllegraEra)
-import Cardano.Ledger.Alonzo (Alonzo, AlonzoEra)
-import Cardano.Ledger.Babbage (Babbage, BabbageEra)
-import Cardano.Ledger.Conway (Conway, ConwayEra)
+import Cardano.Ledger.Allegra (AllegraEra)
+import Cardano.Ledger.Alonzo (AlonzoEra)
+import Cardano.Ledger.Babbage (BabbageEra)
+import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Core (
   AtLeastEra,
   AtMostEra,
@@ -72,9 +69,10 @@ import Cardano.Ledger.Core (
   eraProtVerHigh,
   eraProtVerLow,
  )
-import Cardano.Ledger.Mary (Mary, MaryEra)
-import Cardano.Ledger.Shelley (Shelley, ShelleyEra)
+import Cardano.Ledger.Dijkstra (DijkstraEra)
+import Cardano.Ledger.Mary (MaryEra)
+import Cardano.Ledger.Shelley (ShelleyEra)
 
 -- | Sometimes it is useful to specify that a type corresponds to a latest era that is
 -- currently implemented
-type LatestKnownEra = ConwayEra
+type LatestKnownEra = DijkstraEra