+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Transition.hs
@@ -13,10 +13,11 @@ module Cardano.Ledger.Api.Transition (
 
   -- * Genesis
   EraGenesis (..),
+  NoGenesis (..),
   ShelleyGenesis (..),
   AlonzoGenesis (..),
   ConwayGenesis (..),
-  NoGenesis (..),
+  DijkstraGenesis (..),
 
   -- * Functions for Testing
   tcInitialPParamsG,
@@ -29,6 +30,7 @@ module Cardano.Ledger.Api.Transition (
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis (..))
 import Cardano.Ledger.Api.Era (LatestKnownEra)
 import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
+import Cardano.Ledger.Dijkstra.Genesis (DijkstraGenesis (..))
 import Cardano.Ledger.Genesis (EraGenesis (..), NoGenesis (..))
 import Cardano.Ledger.Shelley.Genesis (ShelleyGenesis (..))
 import Cardano.Ledger.Shelley.Transition (
@@ -53,3 +55,4 @@ mkLatestTransitionConfig shelleyGenesis alonzoGenesis conwayGenesis =
     & mkTransitionConfig alonzoGenesis
     & mkTransitionConfig NoGenesis
     & mkTransitionConfig conwayGenesis
+    & mkTransitionConfig DijkstraGenesis