+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/Tx.hs
@@ -20,8 +20,8 @@ import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import Lens.Micro
 import Numeric.Natural
+import Test.Cardano.Ledger.Api.Arbitrary ()
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Conway.Arbitrary ()
 import Test.Cardano.Ledger.Core.KeyPair (ByronKeyPair (..), KeyPair (..), mkWitnessVKey)
 
 txSpec ::
@@ -104,3 +104,4 @@ spec = do
   txSpec @AlonzoEra
   txSpec @BabbageEra
   txSpec @ConwayEra
+  txSpec @DijkstraEra