+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Arbitrary.hs
@@ -4,6 +4,7 @@ module Test.Cardano.Ledger.Api.Arbitrary () where
 
 import Cardano.Ledger.Api.State.Query (MemberStatus)
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Dijkstra.Arbitrary ()
 
 instance Arbitrary MemberStatus where
   arbitrary = arbitraryBoundedEnum